Refactoring Server Startup to be more idiomatic for Twisted.
