+++ evennia/server/tests/test_amp_connection.py
@@ -11,9 +11,14 @@ from model_mommy import mommy
 from twisted.internet.base import DelayedCall
 from twisted.trial.unittest import TestCase as TwistedTestCase
 
+import evennia
 from evennia.server import amp_client, server, serversession, session
 from evennia.server.portal import amp, amp_server, portal
 from evennia.utils import create
+from evennia.server.service import EvenniaServerService
+from evennia.server.portal.service import EvenniaPortalService
+from evennia.server.sessionhandler import ServerSessionHandler
+from evennia.server.portal.portalsessionhandler import PortalSessionHandler
 
 DelayedCall.debug = True
 
@@ -24,22 +29,23 @@ class _TestAMP(TwistedTestCase):
     def setUp(self):
         super().setUp()
         self.account = mommy.make("accounts.AccountDB", id=1)
-        self.server = server.Evennia(MagicMock())
-        self.server.sessions.data_in = MagicMock()
-        self.server.sessions.data_out = MagicMock()
+        self.server = EvenniaServerService()
+        evennia.SERVER_SESSION_HANDLER = ServerSessionHandler()
+        evennia.SERVER_SESSION_HANDLER.data_in = MagicMock()
+        evennia.SERVER_SESSION_HANDLER.data_out = MagicMock()
         self.amp_client_factory = amp_client.AMPClientFactory(self.server)
         self.amp_client = self.amp_client_factory.buildProtocol("127.0.0.1")
         self.session = MagicMock()  # serversession.ServerSession()
         self.session.sessid = 1
-        self.server.sessions[1] = self.session
+        evennia.SERVER_SESSION_HANDLER[1] = self.session
 
-        self.portal = portal.Portal(MagicMock())
-        self.portal.maintenance_task.stop()
+        self.portal = EvenniaPortalService()
         self.portalsession = session.Session()
         self.portalsession.sessid = 1
-        self.portal.sessions[1] = self.portalsession
-        self.portal.sessions.data_in = MagicMock()
-        self.portal.sessions.data_out = MagicMock()
+        evennia.PORTAL_SESSION_HANDLER = PortalSessionHandler()
+        evennia.PORTAL_SESSION_HANDLER[1] = self.portalsession
+        evennia.PORTAL_SESSION_HANDLER.data_in = MagicMock()
+        evennia.PORTAL_SESSION_HANDLER.data_out = MagicMock()
         self.amp_server_factory = amp_server.AMPServerFactory(self.portal)
         self.amp_server = self.amp_server_factory.buildProtocol("127.0.0.1")
