+++ evennia/server/tests/testrunner.py
@@ -5,8 +5,6 @@ all over the code base and runs them.
 Runs as part of the Evennia's test suite with 'evennia test evennia"
 
 """
-from unittest import mock
-
 from django.test.runner import DiscoverRunner
 
 
@@ -19,16 +17,7 @@ class EvenniaTestSuiteRunner(DiscoverRunner):
     """
 
     def setup_test_environment(self, **kwargs):
-        # the portal looping call starts before the unit-test suite so we
-        # can't mock it - instead we stop it before starting the test - otherwise
-        # we'd get unclean reactor errors across test boundaries.
-        from evennia.server.portal.portal import PORTAL
-
-        PORTAL.maintenance_task.stop()
-
-        # initialize evennia itself
         import evennia
-
         evennia._init()
 
         from django.conf import settings
